+++ src/components/DownloadReports/DownloadReports.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Col } from 'react-bootstrap'
+import Col from 'react-bootstrap/Col'
 import HelpIcon from 'src/components/HelpIcon/HelpIcon'
