+++ src/components/MetricsCounter/MetricsCounter.tsx
@@ -1,9 +1,9 @@
+import React from 'react'
 import { faQuoteLeft, faDownload } from '@fortawesome/free-solid-svg-icons'
 import { faEye } from '@fortawesome/free-regular-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import React from 'react'
 import { Work } from 'src/data/types'
-import { pluralize } from '../../utils/helpers'
+import { pluralize } from 'src/utils/helpers'
 
 
 type Props = {
