+++ src/components/ShareLinks/ShareButtons.tsx
@@ -1,0 +1,10 @@
+'use client'
+
+import {
+  EmailShareButton,
+  FacebookShareButton,
+  TwitterShareButton
+} from 'react-share'
+
+
+export { EmailShareButton, FacebookShareButton, TwitterShareButton }
