+++ src/components/Title/Title.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
-import { Row, Col } from 'react-bootstrap'
+import Row from 'react-bootstrap/Row'
+import Col from 'react-bootstrap/Col'
 
 import styles from './Title.module.scss'
 import { Rights } from 'src/data/types'
